--- index.html
@@ -1,9 +1,0 @@
-<html>
-	<link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.0-beta1/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-0evHe/X+R7YkIZDRvuzKMRqM+OrBnVFBL6DOitfPri4tjfHxaWutUpFmBp4vmVor" crossorigin="anonymous">
-	<body>
-		<h1 class="text-center" style="right:30vw; line-height:5.5em"> Hello Buddies!!!!
-		<div>
-			<img class="rounded mx-auto d-block" style="width:30vw,height:30vw" src="https://i2.wp.com/codigoespagueti.com/wp-content/uploads/2021/11/Este-increible-cosplay-de-Misato-Katsuragi-ha-confirmado-que-es-la-waifu-mas-bella-de-toda-la-serie.png?resize=1280%2C720&quality=80&ssl=1">
-		</div>       
-    </body>
-</html>
