+++ main.py
@@ -1,12 +1,17 @@
-from fastapi import FastAPI
+from fastapi import FastAPI, Request
 from fastapi.responses import HTMLResponse
+from fastapi.staticfiles import StaticFiles
+from fastapi.templating import Jinja2Templates
 
 app = FastAPI()
 
-def generate_html():
-    html = open("index.html").read()
-    return HTMLResponse(html,status_code=200)
+app.mount("/static", StaticFiles(directory="static"), name="static")
 
-@app.get("/",response_class=HTMLResponse)
-async def root():
-    return generate_html()
+templates = Jinja2Templates(directory="templates")
+
+@app.get('/')
+async def index(request: Request):
+  context = {
+    "request": request,
+  }
+  return templates.TemplateResponse("index.html", context)
