+++ templates/categorias.html
@@ -1,0 +1,13 @@
+<!-- Categorías -->
+<div class="space-y-3 lg:w-1/5 lg:px-2 lg:space-y-4">
+  <h3 class="text-lg">Categorías</h3>
+  <a href="#" class="block font-medium text-gray-500 dark:text-gray-300 hover:underline">Ropa</a>
+  <a href="#" class="block font-medium text-gray-500 dark:text-gray-300 hover:underline">Tecnología</a>
+  <a href="#" class="block font-medium text-gray-500 dark:text-gray-300 hover:underline">Juguetes</a>
+  <a href="#" class="block font-medium text-gray-500 dark:text-gray-300 hover:underline">Belleza</a>
+  <a href="#" class="block font-medium text-gray-500 dark:text-gray-300 hover:underline">Deportes</a>
+  <a href="#" class="block font-medium text-gray-500 dark:text-gray-300 hover:underline">Hogar</a>
+  <a href="#" class="block font-medium text-gray-500 dark:text-gray-300 hover:underline">Música</a>
+  <a href="#" class="block font-medium text-gray-500 dark:text-gray-300 hover:underline">Libros</a>
+  <a href="#" class="block font-medium text-gray-500 dark:text-gray-300 hover:underline">Herramientas</a>
+</div>
