+++ templates/index.html
@@ -10,6 +10,8 @@
   <script src="https://unpkg.com/axios/dist/axios.min.js"></script>
   <script src="https://cdn.tailwindcss.com"></script>
   <link href="{{ url_for('static', path='css/styles.css') }}" rel="stylesheet">
+
+  <link rel="icon" type="image/x-icon" href="{{ url_for('static', path='/images/favicon.ico') }}">
 </head>
 
 <body>
