+++ templates/navegacion.html
@@ -11,14 +11,14 @@
     <div class="w-full block flex-grow lg:flex lg:items-center lg:w-auto">
       <div class="text-sm lg:flex-grow text-center">
         
-        <a href="#responsive-header" class="block mt-4 lg:inline-block lg:mt-0 text-white hover:text-blue-200 text-lg mr-4">
-          Item 1
+        <a href="#responsive-header" class="block mt-4 lg:inline-block lg:mt-0 text-white hover:text-blue-200  hover:underline text-lg mr-8">
+          Categoría
         </a>
-        <a href="#responsive-header" class="block mt-4 lg:inline-block lg:mt-0 text-white hover:text-blue-200 text-lg">
-          Item 2
+        <a href="#responsive-header" class="block mt-4 lg:inline-block lg:mt-0 text-white hover:text-blue-200  hover:underline text-lg">
+          Productos
         </a>
-        <a href="#responsive-header" class="block mt-4 lg:inline-block lg:mt-0 text-white hover:text-blue-200 text-lg ml-4">
-          Item 3
+        <a href="#responsive-header" class="block mt-4 lg:inline-block lg:mt-0 text-white hover:text-blue-200  hover:underline text-lg ml-8">
+          Carro
         </a>
       </div>
       <div>
